Rename `constants.js` to `shared.js` and use for helpers
