--- src/blocks/posts-list/constants.js
@@ -1,4 +1,0 @@
-export const MIN_EXCERPT_LENGTH = 10;
-export const MAX_EXCERPT_LENGTH = 100;
-export const MAX_POSTS_COLUMNS = 6;
-export const TERMS_LIST_QUERY = { per_page: -1 };
